separate `MuxTracer`
